se pone la ubicacion en un componente nuevo y se agrega buscar las coordenadas por una direccion escrita
