+++ src/app/app.scss
@@ -1,10 +1,10 @@
 @import "../theme/variables.scss";
 
 ion-app.md{
-    background-color: transparent !important;
-    background-repeat: no-repeat;
-    background-position: center;
-    background-size: 70%;
+  background-color: transparent !important;
+  background-repeat: no-repeat;
+  background-position: center;
+  background-size: 70%;
 }
 
 .list-md {
@@ -12,220 +12,238 @@ ion-app.md{
 }
 
 ion-app, ion-content, .nav-decor {
-    background-color: transparent !important;
-    background-image: none !important; 
+  background-color: transparent !important;
+  background-image: none !important;
 }
 
 body{
-    background-color: transparent !important;
-    background-image: none !important; 
+  background-color: transparent !important;
+  background-image: none !important;
 }
 
 
 
 .flex{
-    &-v{
-        display: flex;
-        flex-direction: column
-    }
-    &-h{
-        display: flex;
-        flex-direction: row
-    }
+  &-v{
+    display: flex;
+    flex-direction: column
+  }
+  &-h{
+    display: flex;
+    flex-direction: row
+  }
+  &-10{
+    flex: 10%;
+  }
+  &-20{
+    flex: 20%;
+  }
+  &-30{
+    flex: 30%;
+  }
+  &-40{
+    flex: 40%;
+  }
+  &-50{
+    flex: 50%;
+  }
+  &-80{
+    flex: 80%;
+  }
 }
 
 .height{
-    &-10{
-        height: 10%;
-    }
-    &-20{
-        height: 20%;
-    }
-    &-30{
-        height: 30%;
-    }
-    &-40{
-        height: 40%;
-    }
-    &-50{
-        height: 50%;
-    }
-    &-60{
-        height: 60%;
-    }
-    &-70{
-        height: 70%;
-    }
-    &-80{
-        height: 80%;
-    }
-    &-90{
-        height: 90%;
-    }
-    &-100{
-        height: 100%;
-    }
+  &-10{
+    height: 10%;
+  }
+  &-20{
+    height: 20%;
+  }
+  &-30{
+    height: 30%;
+  }
+  &-40{
+    height: 40%;
+  }
+  &-50{
+    height: 50%;
+  }
+  &-60{
+    height: 60%;
+  }
+  &-70{
+    height: 70%;
+  }
+  &-80{
+    height: 80%;
+  }
+  &-90{
+    height: 90%;
+  }
+  &-100{
+    height: 100%;
+  }
 }
 
 .width{
-    &-10{
-        width: 10%;
-    }
-    &-20{
-        width: 20%;
-    }
-    &-30{
-        width: 30%;
-    }
-    &-35{
-        width: 35%;
-    }
-    &-40{
-        width: 40%;
-    }
-    &-50{
-        width: 50%;
-    }
-    &-60{
-        width: 60%;
-    }
-    &-70{
-        width: 70%;
-    }
-    &-80{
-        width: 80%;
-    }
-    &-90{
-        width: 90%;
-    }
-    &-100{
-        width: 100%;
-    }
+  &-10{
+    width: 10%;
+  }
+  &-20{
+    width: 20%;
+  }
+  &-30{
+    width: 30%;
+  }
+  &-35{
+    width: 35%;
+  }
+  &-40{
+    width: 40%;
+  }
+  &-50{
+    width: 50%;
+  }
+  &-60{
+    width: 60%;
+  }
+  &-70{
+    width: 70%;
+  }
+  &-80{
+    width: 80%;
+  }
+  &-90{
+    width: 90%;
+  }
+  &-100{
+    width: 100%;
+  }
 }
 
 .center-horizontal{
-    display: flex;
-    justify-content: center;
+  display: flex;
+  justify-content: center;
 }
 
 .center-vertical{
-    display: flex;
-    align-items: center;
+  display: flex;
+  align-items: center;
 }
 
 .end-horizontal{
-    display: flex;
-    justify-content: flex-end
+  display: flex;
+  justify-content: flex-end
 }
 
 .first-item{
-    margin-top: 1em;
+  margin-top: 1em;
 }
 
 .last-item{
-    margin-bottom:1em;
+  margin-bottom:1em;
 }
 
 .margin-top-bottom{
-    margin-top: 0.5em;
-    margin-bottom: 0.5em;
+  margin-top: 0.5em;
+  margin-bottom: 0.5em;
 }
 
 .confirm-alert{
-    > div{
-      background-color:$first-detail-color!important;
-      text-align: center;
-      font-weight: bold;
-    }
-    .alert-message{
-        color: #ffffff;
-        font-size: 1.3em;
-        padding: 1em;
-    }
-    .alert-head{
-        color: #ffffff !important;
-    }
-    .alert-button{
-        color: #ffffff;
-    }
+  > div{
+    background-color:$first-detail-color!important;
+    text-align: center;
+    font-weight: bold;
+  }
+  .alert-message{
+    color: #ffffff;
+    font-size: 1.3em;
+    padding: 1em;
+  }
+  .alert-head{
+    color: #ffffff !important;
+  }
+  .alert-button{
+    color: #ffffff;
+  }
 }
 
 .success-alert{
-    text-align: center;
-    > div{
-        background-color: $first-detail-color !important;
-    }
+  text-align: center;
+  > div{
+    background-color: $first-detail-color !important;
+  }
 }
 
 .error-alert{
-    text-align: center;
-    > div{
-        background-color: $error-color !important;//$error-color !important;
-    }
+  text-align: center;
+  > div{
+    background-color: $error-color !important;//$error-color !important;
+  }
 }
 
 
 .item-md {
-    border-radius: 12px;
+  border-radius: 12px;
 }
 
 .item-input .label-md{
-    margin-top: 5px;
+  margin-top: 5px;
 }
 
 .select-md .label-md{
-    margin-top: 5px;
+  margin-top: 5px;
 }
 
 .item-label-floating .select-md{
-    margin-top: 5px;
-    margin-bottom: 5px;
+  margin-top: 5px;
+  margin-bottom: 5px;
 }
 
 .item-label-floating .text-input-md{
-    margin-top: 5px;
-    margin-bottom: 5px;
+  margin-top: 5px;
+  margin-bottom: 5px;
 }
 
 .fila-image-container{
-    height: 100%;
-    width: 65px;
-    position: relative;
-    display: flex;
-    justify-content: center;
-    align-items: center;  
+  height: 100%;
+  width: 65px;
+  position: relative;
+  display: flex;
+  justify-content: center;
+  align-items: center;
 }
 .fila-image{
-    max-width: 100%;
-    padding: 0.1em;
-    position: absolute;
-    display: block;
-    max-height: 100%;
+  max-width: 100%;
+  padding: 0.1em;
+  position: absolute;
+  display: block;
+  max-height: 100%;
 }
 
 .flex-container{
-    padding: 0;
-    margin: 0;
-    list-style: none;
-    -ms-box-orient: horizontal;
-    display: -webkit-box;
-    display: -moz-box;
-    display: -ms-flexbox;
-    display: -moz-flex;
-    display: -webkit-flex;
-    display: flex;
-    flex-direction: column;
+  padding: 0;
+  margin: 0;
+  list-style: none;
+  -ms-box-orient: horizontal;
+  display: -webkit-box;
+  display: -moz-box;
+  display: -ms-flexbox;
+  display: -moz-flex;
+  display: -webkit-flex;
+  display: flex;
+  flex-direction: column;
 }
 
 .flex-item {
-    background: rgba(49, 49, 49, 0.96);
-    width: 100%;
-    height: 90px;
-    margin: 0.2em;
-    line-height: 100px;
-    color: white;
-    font-weight: bold;
-    font-size: 2em;
-    text-align: center;
-    padding: 0.2em;
+  background: rgba(49, 49, 49, 0.96);
+  width: 100%;
+  height: 90px;
+  margin: 0.2em;
+  line-height: 100px;
+  color: white;
+  font-weight: bold;
+  font-size: 2em;
+  text-align: center;
+  padding: 0.2em;
 }
