+++ src/pages/alta-pedido/alta-pedido.html
@@ -44,22 +44,8 @@
       </ion-col>
     </ion-row>
   </ion-grid>
-  <ion-row>
-    <ion-col>
-      <ion-item>
-        <ion-label floating>Ingresar direccion</ion-label>
-        <ion-input type="text" required name="origen" [(ngModel)]="direccion"></ion-input>
-      </ion-item>
-    </ion-col>
-    <ion-col>
+  <page-ubicacion [direccion]="direccion"></page-ubicacion>
 
-    </ion-col>
-  </ion-row>
-  <ion-row style="width: 100%; height: 60%;">
-    <ion-col>
-      <div #map id="map"></div>
-    </ion-col>
-  </ion-row>
 
 </ion-content>
 <ion-footer (click)="Confirmar()">
