+++ src/pages/alta-pedido/alta-pedido.scss
@@ -4,11 +4,6 @@ page-alta-pedido {
         background-color: $first-detail-color;
      }
 
-    #map {
-      width: 100%;
-      height: 100%;
-    }
-
      .fle {
         width: 100%;
         height: auto;
