+++ src/pages/alta-pedido/alta-pedido.ts
@@ -19,18 +19,6 @@ import { ParamsService } from '../../services/params.service';
 })
 export class AltaPedidoPage {
 
-  //Google map tutorial: https://www.joshmorony.com/ionic-2-how-to-use-google-maps-geolocation-video-tutorial/
-  //Navigation geolocation: https://www.joshmorony.com/implementing-turn-by-turn-navigation-with-google-maps-in-ionic/
-  //Apikey name: appcomandapps / AIzaSyBJffx9HcTzjsw4mBAEMti_BUuXA9VTIaA
-  //Cloud de servicios de google apis (mail de ger) https://console.cloud.google.com/google/maps-apis/overview?onboard=true&project=appcomandapps-1541641825014&consoleUI=CLOUD
-
-  //Ejemplos de proyectos:https://www.joshmorony.com/shhh/?convertkit=true&first_name=&email=carina.amuz%40gmail.com
-
-  @ViewChild('map') mapElement: ElementRef;
-  @ViewChild('directionsPanel') directionsPanel: ElementRef;
-  map: any;
-  marker:any;
-  direccion:any;
   bebidas:any;
   comidas:any;
   reservas:any;
@@ -43,6 +31,7 @@ export class AltaPedidoPage {
   user:any;
   beb = 0;
   pla = 0;
+  direccion:any = {value:""};
 
 
   constructor(public navCtrl: NavController,
@@ -54,14 +43,6 @@ export class AltaPedidoPage {
               public geolocation: Geolocation,
               private geocodingProvider: GeocodingProvider,
               private platform: Platform) {
-    this.platform.ready().then(()=>{
-      this.geolocation.getCurrentPosition({ timeout: 30000, enableHighAccuracy:false })
-        .then(data => {
-          this.cargarMapa(data.coords.latitude, data.coords.longitude);
-        });
-      //Datos solo para testing si no anda el currentposition en el browser descomentar esta linea
-     // this.cargarMapa(-34.6344243, -58.386632199999994)
-    });
 
     this.navParams.get("reserva") ? this.reservaKey = this.navParams.get("reserva") : null;
     this.navParams.get("dniCliente") ? this.reservadniCliente = this.navParams.get("dniCliente") : null;
@@ -224,82 +205,9 @@ export class AltaPedidoPage {
     }
   }
 
-  mapClick(this, event){
-    var latLng = event.latLng;
-    // this.map = new google.maps.Map(this.mapElement.nativeElement, mapOptions);
-    this.obtenerDireccionPorCoordenadas(latLng.lat(), latLng.lng());
-    this.addMarker(latLng.lat(), latLng.lng());
-  }
 
-  private cargarMapa(lat, long) {
-    let latLng = new google.maps.LatLng(lat, long);
-    let mapOptions = {
-      center: latLng,
-      zoom: 15,
-      mapTypeId: google.maps.MapTypeId.ROADMAP
-    };
-    this.map = new google.maps.Map(this.mapElement.nativeElement, mapOptions);
-    google.maps.event.addListener(this.map, 'click', event => {
-      this.mapClick(event)
-    });
-    this.obtenerDireccionPorCoordenadas(lat, long);
-    this.addMarker(lat, long);
-  }
 
-  private obtenerDireccionPorCoordenadas(lat, long){
-    this.geocodingProvider.obtenerDireccion(lat, long).then(response =>{
-      this.direccion = response;
-    }, error => {
-      console.log(error);
-    });
-  }
 
-  private addMarker(lat, long){
-    let latLng = new google.maps.LatLng(lat, long);
-    //Si existe otro punto en el mapa lo elimina
-    if(this.marker){
-      this.marker.setMap(null);
-    }
-    this.marker = new google.maps.Marker({
-      map: this.map,
-      animation: google.maps.Animation.DROP,
-      position: latLng
-    });
 
-    let content = "<h5>" + this.direccion + "</h5>";
-
-    this.addInfoWindow(this.marker, content);
-  }
-
-  private addInfoWindow(marker, content){
-    let infoWindow = new google.maps.InfoWindow({
-      content: content
-    });
-    google.maps.event.addListener(this.marker, 'click', () => {
-      infoWindow.open(this.map, marker);
-    });
-  }
-
-  //este sirve para poner dos rutas de destino y te indica que camino tomar
-  startNavigating(){
-    let directionsService = new google.maps.DirectionsService;
-    let directionsDisplay = new google.maps.DirectionsRenderer;
-
-    directionsDisplay.setMap(this.map);
-    directionsDisplay.setPanel(this.directionsPanel.nativeElement);
-
-    directionsService.route({
-      origin: 'adelaide',
-      destination: 'adelaide oval',
-      travelMode: google.maps.TravelMode['DRIVING']
-    }, (res, status) => {
-
-      if(status == google.maps.DirectionsStatus.OK){
-        directionsDisplay.setDirections(res);
-      } else {
-        console.warn(status);
-      }
-    });
-  }
 
 }
