+++ src/pages/ubicacion/ubicacion.scss
@@ -1,0 +1,10 @@
+@import "../../theme/variables.scss";
+
+page-ubicacion{
+
+  #map {
+    width: 100%;
+    height: 100%;
+  }
+
+}
